api/requirements.txt
asgiref==3.11.1
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
inflection==0.5.1
pip==25.3
PyYAML==6.0.3
sqlparse==0.5.5
tzdata==2026.1
uritemplate==4.2.0